Fix CommandProcessor queue processing
It was reported on the mailing list that commands queued in a
non-periodic activity were not all dequeued and only the first
item was processed. This patch makes the CommandProcessor to
process all queued commands in a step() and adds a unit test
for checking this behaviour.

From: Peter Soetens <peter.soetens@fmtc.be>

git-svn-id: https://svn.mech.kuleuven.be/repos/orocos/trunk/rtt@29848 ce417995-dfc9-0310-95a0-acdaff106893