build: update corba-specific build variable
first, the RTT_CORBA_CFLAGS should be used instead of RTT_CFLAGS for
corba-specific parameters. Moreover, there is still a problem with
OmniORB, for which the use of IDL-generated headers still require
the specification of rtt/corba as the include directory. Therefore,
set it in the .pc file.

From: Sylvain Joyeux <sylvain.joyeux@dfki.de>

git-svn-id: https://svn.mech.kuleuven.be/repos/orocos/trunk/rtt@29917 ce417995-dfc9-0310-95a0-acdaff106893