start the redesign of the data flow implementation
In this new implementation, data flow is represented by mainly
two components:
  * the ports are the dataflow endpoints at the task level. They
    are represented in practice by WritePort<T> and ReadPort<T> objects.
  * the connection is the data path between two ports. There can be
    multiple connections going out of a write port, but only one
    connection entering a read port.

Connections are established from the WritePort side. ConnElement
are managed as shared pointers (more specifically, intrusive_ptr)
*from the writer* (i.e. a connection element holds an intrusive_ptr
to the next element in the chain, but a normal pointer to the
previous element). For now, there is no mt-guarantees whatsoever.

A connection is uniquely defined by its writer port, its read port and
a connection policy. This later object, of class ConnPolicy, defines
how the connection should behave. For now, we support the type of
data holder (data or buffer), and the type of synchronization (locked
or lock-free).

The connections are made of connection elements (ConnElement) objects.
A connection is built in two halves. If the connected endpoints are
in the same process, that concept is useless. However, if the two
are on different machines, it takes its meaning:
  * the reader half is the part that is built in the process
    where the ReadPort is located.
  * the writer half is the part that is built in the process
    where the WritePort is located.

In practice, it allows to have different policies that dictates
different ConnElement structure.