make build(Reader|Writer)Half independendent of resp. writer/reader
Until now, buildReaderHalf and buildWriterHalf were getting both
the reader and writer ports to set up the connection. This is a
hassle in practice (because we need to have access to both port
objects) and IMO unnecessary.

This commit therefore adds the constraint that building one half
of the connection is independent of the particular remote port.
Additional setup can still be done at the connection establishment
level (i.e. createConnection by default, or for instance a specific
connection engine).