clarify the semantics of write() and read() return values for both connections and ports
Right now, the semantic of Connection<T>::write was really unclear.
This commit clarifies it. One consequence of this analysis is that
there is no well-defined meaning for WritePort<>::write. Therefore,
this method now returns void.

The meaning for Connection<T>::write is now "dangling connection", as
for instance a socket connection for which the other side is closed.
WritePort<T>::write reads that return value and automatically deletes
these dangling connections.