corba: adapt the CORBA transport to the new dataflow interface
This commit makes the CORBA transport work with the new data flow interface. The
cornerstone of this is still the DataFlow IDL interface, which provides the
buildReaderHalf service.

Note about RemoteConnElement<T>: this class is both a ConnElement and a
CORBA refcounted servant. We therefore make both sides use the same refcount,
but taking care that the refcounting starts at 1 as CORBA dictates it.

Note about ControlTaskProxy: the changes brought by the new data flow interface
have removed the CorbaPort class -- which should come back at some point. For
now, simply comment out that part in ControlTaskProxy::synchronize().