make the new dataflow connection interface MT-safe
This commits uses ListLockFree<T> for connection management in
WritePort<T>, making the connection management interface
(createConnection and disconnect) MT-safe.

Moreover, some fixes have been implemented in the implementation
of connection elements. Mainly, we must make sure that every
method which is using a pointer to ConnElementBase should also
be holding a local shared_ptr on that connection element.