corba: implement local proxies for remote ports
This commit implements RemoteReadPort and RemoteWritePort which
are local CORBA proxies for remote ports. In this implementation,
a remote port is represented by a CORBA client to a DataFlowInterface
and a name.

Because of this choice, some specific operations of WritePortInterface
are not yet implemented on RemoteWritePort. This is because it is
impossible to get a DataFlowInterface object from a given port
(and therefore, for instance, RemoteWritePort::createConnection
cannot be implemented). To fix this, one would have to have a CORBA
representation of the port objects themselves.

To allow transparent use of Remote*Port in connections,
I had to change the way connections are
represented in WritePort<T>. Before, a PortInterface* pointer
was representing the read port of the connection, and used
later to disconnect a specific port. Now, an opaque PortID
type is now used. That PortID is generated by the virtual
PortInterface::getPortID and equivalence is tested with
PortInterface::isSameID. This PortID object is owned by
WritePort<T> and deleted by it on disconnection.

That way, it is possible to have temporary PortInterface objects
(like on-stack RemoteReadPort) used to create permanent connections
*and* still be able to disconnect selectively later on.

Finally, I defined the semantic of PortInterface::clone
and PortInterface::antiClone, so that it can be defined
(hopefully) meaningfully on remote ports. The semantic is now
that these methods must return local ports that have the same
type and name than the remote ones. For instance, a RemoteWritePort
representing a remote WritePort<double> would clone into
WritePort<double> and antiClone into ReadPort<double>.