add ReadPortInterface::getDataSource()
This is a useful interface for uniform access to data-reading
elements, for instance in the task browser and friends.

This commit adds the interface to read ports. To access write
ports, one should first create the read port with antiClone(),
connect with createConnection and create the data source. This
is so because we cannot easily implement the memory management
needed to implement WritePortInterface::getDataSource