have ChannelElement::signal have the identical return semantics than write
This commit changes the interface of ChannelElement::signal()
so that it returns a boolean. The semantics of that return value
is the same than for write().

Specifically, a false return value indicates that the underlying
communication layer requests that the channel gets removed, for
instance because of communication errors.

In no way does it put a requirement on the reliability of the layer.
A return value of true does *not* indicate that the signal (or
the sample in case of write) got through.