implement a status() method for commands
The problem with the current interface is that the actual
command state is very hard to get, as it requires a
deep knowledge of how things are getting.

Moreover, on the CORBA layer side, it requires multiple calls.

The new interface provides directly the command's state.
Moreover, it helps getting things streamlined by making
sure each command implementation follows the command state's
specification.