corba: in CORBA's createConnection, don't
This side of the fence is to be used by a CORBA-aware client,
which should therefore be able to interpret and act on CORBA
exceptions.

We should therefore not filter out corba exceptions, as they
are useful to our CORBA client. Note that it is different from
8ae41a09f23ca8e385d5e41c7ff6ccaf77dc214c, where the CORBA
exceptions where leaking to non-CORBA user code.