including a smart pointer for explicit data sharing semantics
The goal of this pointer is to allow efficient passing of big data structures
through ports. Right now, that would require at least 3 copies of the data. With
this pointer, one can push a sample on a port, and have the readers create a
copy only when
 - they need it
 - there is at least one other component that could use the given sample

That would be very useful for instance in the context of image processing.