This patch adds initialisation of the connection using a data sample and checking of connection validity for out-of-band connections.
The idea is that when a chain of channel elements is created, the
output port is asked to send a data 'example' through the connection
and the input port is asked to check if its channel elements are ready.

Because of these two 'messages' going in both directions, we can initialise
and validate both ends of the data connection for real-time communication.
The data sample allows the channel elements to allocate enough memory
to hold its (dynamic) size, while the input check allows out-of-band
transports to setup their data receiving (knowing that output is up
and sent a sample).

* Storing the last output value in an output port is now turned on by default.
* Initializing a data connection with the last output value of the port
  is now turned on by default.