transports: Remove the *blob functions from DataSource and relocate them in the proper transport object.
There was no need to clutter the DataSource interface with the 'blob'
functions for converting data to transportable objects and back.
The create/updateBlob functions have been renamed to createAny/updateFromAny
for corba (which really works with anys), while keeping their name in
mqueue transports (which really works with blobs).