dataflow: Allow multi-connection inputs.
This factors out the connection management into a separate class.
This allows both input and output ports to use connection management.

Also the DataFlow IDL interface had to be adapted to support this
mode of operation and many other changes were required in the data flow
classes.

It seems the PortID will need to replaced by ConnectionID to allow
streams to be managed as well.