events: Ripped out asynchronous event handling and EventService
This patch removes any occurence of asynchronous RTT::Event handling code.
Some locations could not be fixed without a MessageProcessor, so these have the
code removed and an assert(false) put in. Broken are:
 * newDataOnPortEvent
 * event processing in StateMachine
 * state machine transition event parsing

The new primitive to use in place is RTT::internal::Signal for in-thread
function dispatching.

Some Event code is still left in because it will be re-used for integrating Signal
with the Method/Operation interface. These classes will be gradually removed/renamed
or refactored.