methods: Implemented asynchronous collect functions.
An asynchronous Method invocation requires the ExecutionEngine
of the owner/executor and of the caller, for receiving the callback.

This patch implements call/send/collect/collectIfDone/ret for
arity 0->4 + return value.

This is only the basic C++ mechanism, we further need the Operation
class, new registration code and support in scripting and corba.
Also there is not yet support for a '3rd party' thread/EE that
executes the method in case neither component nor client wishes
to do so. Most of this stuff all belongs in the new Operation class.