methods: Remove weak_ptr from DS method implementations.
Since shared_ptr is natively supported by the boost::bind/invoke functions and
forcing in weak_ptr was causing bloat for a feature we actually don't want
(only for 1.x BC scripting), this patch standardises in keeping the pointer
to the script in a shared_ptr. Some other means will have to be invented to
force a cleanup.