scripting: Implemented send/collect in scripts
Unit test added in program_test. Hacky in ExpressionParser,
but does what is expected. When a SendHandle is seen in a script,
a dummy is installed in the AttributeRepository, which gets filled
in by the real SendHandle, once the send() invocation has been seen.
sh.collect(args) had to influence parsing of named values, such that
when a keyword was seen (ie '.collect') the text was passed to the
DataCallParser, which handles the collect then.