extras: do not call reset() in write_access()
write_access() checks if we are sole owner, and based on that returns either a
copy or directly the buffer. It called reset() to release the pointer
automatically once we got read/write access to the data.

This behaviour is actually quite counter-intuitive, as repeated calls to
read_write() will return first a data pointer and then NULL. So, change it so
that the call to reset(), if desired, is done explicitely