types: fix copy-on-write race against std::string implementation
I have to write this down in case this bites us again in the future:
* Orocos scripting assumes that a string of size 'N' will not decrease
  its capacity when it is assigned a string of size '< N'.
* Same for a std::vector

The testTypes unit test checks if this assumption is true. Of course,
it isn't in some cases. This patch works around one more such case.
We already worked around asigning two strings by overloading the
ValueDataSource::set() in case of string. Now we also had to overload
it's constructor in case of string.

The problem is that string uses a copy-on-write implementation that
discards previous capacity in favour of the newly written string,
and allocates a capacity equal to that string's size. Hence above
assumption fails. The only way to counter that is to 'pre-allocate'
a new string for *each* string in scripting such that it is certainly
not a copy-on-write (cow) target. Since we use 'clone()' a lot in scripting,
each clone() was a possible 'shallow string clone' which could trigger
a future 'cow'. The only work around is to make sure that clone() makes
a deep copy instead of a shallow one of the string. This is done by
passing the string::c_str() in the constructor of ValueDataSource, which
forces a 'deep' copy. The same trick had been applied in ::set() earlier.

In order to have the above work, the buildVariable() implementation for
std::string had to be adapted to fill the 'preallocated' string with
spaces instead of '\0' because the c_str() of '\0' is the empty string,
which is not what we want to use in the constructor of ValueDataSource :/

Conflicts:

	rtt/types/RealTimeTypekitTypes.cpp
	tests/types_test.cpp

Note: conflicts resolved for 1.x branch.

From: Peter Soetens <peter@thesourceworks.com>

git-svn-id: https://svn.mech.kuleuven.be/repos/orocos/trunk/rtt@31978 ce417995-dfc9-0310-95a0-acdaff106893