Fixes race condition in EventProcessor
The args are read and stored in a
temporary variable. Then this is used to get the arguments for the function
call.
Before the args where read for each argument. You could end up with
inconsistent data the the subsequent read's where pre-empted by a call to
DataObjectLockFree.

From: Butch Slayer <butch.slayers@gmail.com>

git-svn-id: https://svn.mech.kuleuven.be/repos/orocos/trunk/rtt@31996 ce417995-dfc9-0310-95a0-acdaff106893