FDActivity: do not assume that NonPeriodicActivity::stop() waits for loop() to return
If stop() is called from within a hook (as for instance because fatal() is
called in updateHook()), then NonPeriodicActivity::stop() will NOT wait for
loop() to return.

This was causing an infinite loop as FileDescriptorActivity::stop() was closing
the communication pipe, and then FileDescriptorActivity::loop() never stopped.

Close the pipe in loop() to fix that.