core: fix misuse of grow()
When the connection management was implemented, I thought that grow() was adding
one more element to the lock-free list. It actually does not: it only grows the
expected size of the list.

I think grow() is dangerous, and therefore replaced its use by reserve() with
a total size. This fixes crashes when more than 10 connections were created on
the same port.