operations: add synchronous operation factories
This patch makes it possible to register an operation as synchronous.
The port services use this in order to install a reduced version of
their read/write/last operations, which are only meant for in-component
thread, ie synchronous, usage.

Synchronous operations are not exported throught the CORBA layer.

This solution is here to reduce compilation time of typekits mainly,
and partly because reading/writing ports should only be done by the
component itself.

If we find a better solution, this approach may be deprecated.

In-process asynchronous usage of the operation will still be tolerated,
since it's only the Service interface that is limited, not the Operation
object itself.

Initial patch suggested by Sylvain Joyeux and worked out by Peter Soetens.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>