remote: fix remote method implementation wrt (const) refs.
The pointer to the data can only be filled in after the method
has been invoked. So we require 'Late' data sources that allow
to set this pointer after the data source has been constructed,
but before the target method reads them out.

There is likely a similar issues with return values, which has
not been fixed or tested yet with this patch.