dataflow: manage both input and output with a shared pointer
This fixes a crash-at-disconnection problem.

All channel operations are protected with shared pointers. It means
that when an operation propagates itself along the channel, it always
takes a smart pointer on the next element.

However, the 'input' attribute of ChannelElement is a plain pointer.
In practice, it means that there is a race condition at disconnection
time, between the disconnect operations that deletes an element and
other operations that are trying to get a shared_ptr on it. This
does not exist on the other direction, as we copy a shared_ptr
into a shared_ptr.

So, store the input attribute as a shared_ptr, and make sure that
the disconnect operation resets both pointers to 0.