corba,dataflow: do not explicitely reset remote_side to 0
Since the rest of the channel implementation properly handles
concurrent disconnection + normal operations, so should the
CORBA channel element. Setting the remote_side to 0 is not useful
(the remote_side CORBA proxy will be deleted when the channel
element is), and leads to crashes if a write/read is done as
the same time than a disconnection.