core: make ConnectionManager MT-safe
This is a second try. I first tried to make the original
implementation MT-safe, but had quite a few issues on the way due to
how things were implemented, that could not be solved as-is.

Using lock-free data structures in ConnectionManager was harmful
as the principle behind these data structures is that operations
done one the elements of the structure have to be idempotent,
which is not the case for dataflow operations. One could for
instance have two write() on the same channel happening because
the list is being modified at the same time than the port writes
to it.

Moreover, the current implementation was using an optimization in
case of single connections, which was complexifying logic (and
MT-safety considerations) quite a bit. I considered that it was
therefore a case of early optimization and dropped it, until we
get a rock-solid connection manager.

The new implementation uses std::list and mutex for locking. The
amount of code during which the mutex is taken is reduced to the
strict minimum, to limit interactions between connections/disconnections
and write/read.