operations: fix crash when cleaning up operation caller after send().
In the OperationCallerC/SendHandleC combo, the arguments are stored in DataSources,
which are kept (indirectly) by the OperationCallerC. In case this object is
prematurely destroyed, the operation will segfault because of freed arguments.

The solution is to let SendHandleC also have a refcount to the operation's
data source and block in the destructor until it completed or failed.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>