corba: temporarily lift blocking collect from OperationKeeper
There is still a bug in there which deadlocks the destructor of OperationKeeper.
It blocks in the remote Global EE which never receives a message. Strangely
enough, also the is_executed flag is still false, so the blocking seems
legitimate. So we are checking the wrong flag, or the UnitTest effectively tests a deadlock,
and it's a UT bug.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>