operations: implement setAutoCollect feature
This allows the user of a SendHandleC to determine if that object
should collect upon destruction or that it can safely free resources.
This flag is shared between all copies of the same SendHandleC.

Resources can be cleaned up safely during remote invocations,
but need to be kept until the op completes for local invocations.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>