list: use recursive mutex to allow apply-within-apply
When the RTT is built with OS_NOASM, this list is used by
the StateMachineProcessor, which uses apply both in
step() and in finalize(). As such, a stop() within step()
will call finalize, and the implementation would dead-lock.

The solution is to use a recursive mutex and warn against
faultive use of apply (like calling erase/clear within apply()).

Port from 1.x line.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>

git-svn-id: https://svn.mech.kuleuven.be/repos/orocos/trunk/rtt@33149 ce417995-dfc9-0310-95a0-acdaff106893