plugins: only load libraries ending in .so,.dll,... during dir scan
This was necessary to let the symlinks to their work, ie
point to the most recent version, unless the user overrides
with a specific version (like foo.so.1.2.3) in which case
that version is loaded. So the .so|.dll|.dylib extension
restriction only applies when scanning a whole directory
for plugins or types (RTT_COMPONENT_PATH subdirs etc).

Signed-off-by: Peter Soetens <peter@thesourceworks.com>

git-svn-id: https://svn.mech.kuleuven.be/repos/orocos/trunk/rtt@33156 ce417995-dfc9-0310-95a0-acdaff106893