types: allow to inspect a const struct type
The type introspection can only handle assignable datasources.
We work around this by creating an alias data source to a
copy of the struct's member. The alias makes sure that we
won't write the result (not assignable), while the copy is
safe because it was const in the first place, so the original
won't change.

The only clean solution is to merge AssignableDataSource
with DataSource and have 'assignability' as a property of
DataSource. This would avoid a lot of duplication of code
and passing around of datasources just to mimick the right
behavior.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>