corba: properly deregister channels from channel_list
In some cases, channels get registered in a channel_list list.
However, they were not deregistered when disconnected.

The explicit call to CORBA::release in the new deregisterChannel method call has
been found by trial-and-error, so it might not be the proper fix.  However, it
fixes a memory leak that exists without it, where the value returned by _this()
is not freed.