scripting: unload policy for function graphs and execution engine
It is now mandatory to specify how a function graph must behave in execute(),
with respect to triggering an unload action. program scripts are managed by the
ScriptingService, and should never unload themselves. plain functions are
not managed and must trigger the unload action by returning false in execute().

In addition, the ExecutionEngine was modified in order to call unloading()
in the client thread instead of its own thread in case of a removeFunction
case.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>