scripting: correctly implement stack for runScript
Variables declared on earlier lines were not recognized. This patch
introduces a stack for each run in which these can be stored.
More testing for all kinds of syntax as well, I had to loosen
the function parser in order to get to the statement parsing
without a fatal exception.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>