operations: trim down broken ret_impl for remote operations
The current implementation or ret_impl() couldn't possibly work because
the vStore of sendargs has not been initialized in the RemoteOperationCaller
holding a SendHandleC (only collectargs is). The current code is broken too,
but at least takes less time to compile and less space in memory.

Not sure if this is fixable. The prefered solution would be to remove the ret()
feature altogether, but call() of LocalOperationCaller needs it to return
results of a call on an OwnThread operation (using ret() on the returned
SendHandle). So if there is another way to write LocalOperationCaller, the
feature can be removed.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>