define InputPort<T>::flush
One goal of the port interface is to make the usage of
data or buffers completely transparent to the module.

However, to achieve getting "the most uptodate sample" available
on a port, one would have to do a bit of logic.

flush(sample) implements this logic

It reads all the new samples available and returns the last one or,
if no new sample are available, returns NoData / OldData as expected.