plugins: implement package/project loading
Plugins are now loaded 'a project at a time' or a 'package at a time'.
This means whole subdirectories are loaded, with all plugins in them,
instead of trying to locate individual shared library files. This
allows us to drop the 'heuristics' which were looking for individual
files to load (which are named platform dependently).

Signed-off-by: Peter Soetens <peter@thesourceworks.com>