engine: fix stop() such that it will call breakUpdateHook if necessary.
stop() needs to synchronize with updateHook() in order to have thread-safe execution
of the stopHook() with respect to the updateHook().
The stop() mechanism of the underlying thread is used for this, such that no
secondary synchronization needs to be implemented. I had to relax queue management
in order to allow this running->stopped->running transition, such that queueing of
functions or ops is still possible with a stopped engine.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>