core: fix having OldData always return the same sample than the last NewData
The piece of code that disappeared from the ConnectionManager is
the one that, in select_if, was clearing the current input port
when cur_channel is changed.

This call ensures that there is at most one incoming channel that
returns OldData at all times. Therefore, unless one other channel
returns NewData, one that calls #read repeatedly will always get
the right value. Moreover, NoData will be returned if the
current channel is disconnected -- which is properly the best
behaviour in this case.

Because of this behaviour, select_if is not really a generic method.
Rename to select_reader_channel as it is really what it is meant for.