corba: try remote disconnection if local port disconnection fails
There is an optimization in the CORBA connection handling that
uses normal non-CORBA connections between local ports. A fix
recently made sure that CORBA was able to disconnect local
ports.

However, there are ways to disable that optimization. So, when
disconnecting local ports in CORBA, we must try the normal CORBA
disconnection process if the "local port disconnection" fails.