tests: move the corba server creation/deletion to separate programs
There are two new programs built: setup_corba and cleanup_corba, which
(on UNIX systems) create and kill the corba servers we require for the
ipc tests. This allows you to start the servers manually and run the
tests multiple times, and even allows to run these tests on non-unix
systems.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>