corba: initial implementation of operation caller proxies
We need these in order to do A-B-C connections, where
B connects two remote A and B component interfaces.

This patch adds the OperationCallerProxy object that will forward
all 'setImplementationPart' calls to the remote operation caller's
service requester object which will then do the real setup.

This patch allows you to see the requests of a proxy, but can not
yet make the connection.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>