types: detect double type registration, report it and install an alias
For this to work, the DataSourceTypeInfo handler installation needed
to be delayed and moved into a new virtual function of TypeInfo.

It will alias the old name to the new type and then delete the
old object. There is a little bit bookkeeping going on because of the
double occurence of the type info pointer both globally and in the
TypeInfoRepository data map.

Aliasing required smarter cleanup in the destructor if the TypeInfoRepos.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>