properties: add new composition/decomposition functions
updateProperties() et al. must be rewritten not to do compose/decompose, but
only try to update what is. The compose/decompose decision is to be made
by the caller of updateProperties().

This patch is not ready yet, adds an initial unit test that tests
compose/decompose of an vector<double>

Signed-off-by: Peter Soetens <peter@thesourceworks.com>