avoid non-static global variables
if shared libraries are dl-opened *and* linked "in a normal way", then
the global variables get shared -- as it should -- *but* they also get
deleted as many times as the library got loaded, which leads to multiple
frees/deletes on the same object at shutdown.

This affects symbols defined in the main RTT library, and is therefore
*not* solved by only using the plugin loader.

A workaround is to never have public global variables with destructors.
Global variables should instead be defined static in their compilation
unit and accessed through method calls.

This patch does so for a few problematic ones. It solves crashes at
shutdowns that I have been experiencing.