types: compose the items of a sequence first before composing the sequence itself
The problem with the previous implementation was that it did not
offer the child elements a chance to use their type specific
composition function, since we decomposed the whole sequence and
then did an update. By piece-wise composing each element, we
respect/use the composeType function and the correct code gets
executed.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>