reduce number of sample copies during data transport
This patch changes the logic, how the last sample of
an port is saved. The new logic makes use of the previous
changed BufferInterface. This patch is aimed at reducing the
number of copies in the way of a sample, from a write to a read
operation.

In the previous implementation 5 copies were needed in the Locked and
Lock_Free case (reading until OldData is returned).
With the new implementation there will only be two copies in the Lock_Free
case and three copies in the Locked case.

Testing showed that this reduces the CPU load by ~1/4 on a 8K samples
being send at a kHz.