added ability to choose weather old data should be copied on read operation
The read operation on InputPort and ChannelElement was extended by a new
boolean argument copy_old_data. If this argument is set to true the
original behavior is kept. If copy_old_data is false, the given sample
will only be updated if the FlowStatus is equal to RTT::NewData.

Tests with 8K samples at 1 kHz showed showed a drop of cpu usage by 1/4.

Note, the default argument for InputPort::read is false, so old code
should not be affected by this change.