remove invalid assert in BufferLockFree
A recent change increased the size of the memory
pool. This caused a assert to fail in the case
that the buffer is full and a sample was written.
As the assert relied on the fact, that the buffer
and memory pool size is equal it is wrong and was
removed.