internal: fix enable_shared_from_this wrt weak_ptr API
The weak_ptr in 1.40 contains a method named _empty() which is not present
in older versions. So we can't use it to track initialisation of the weak_ptr.
We use a boolean to store that info. Maybe weak_this_.expired() would have
done too, but since they added _empty() and did not use expired(), there
must be a reason for not using expired().

Signed-off-by: Peter Soetens <peter@thesourceworks.com>