typekit: always provide the 'extern' modifier
We don't use the classical extern trick[1] since Types.inc/Types2.inc don't
use the RTT_TYPEKIT_EXT_TMPL macro *and* we include this header in our own
project to speed up that already (so we need 'extern' in our own project).
It's all quite confusing right now
and not very clean, but it is doing what's supposed to be hapening.

[1] use 'extern' when external sources include the header and omit 'extern'
when we're compiling it ourselves.