compilation errors rtt 2.3.0 boost 1.46
A new boost version has been released, 1.46.
The boost/filesystem lib has had a major revision,
for that small changes need to be made to PluginLoader.cpp

Furthermore, I get errors because 'assertion' is ambiguous in the Parsers.
I suppose it has to do with changes to boost/assertion.
I added namespace specifiers in front of assertion,
not sure though if this is what you think is appropriate.
It made the compilation succeed.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>