signal: fix cleanup/disconnection logic of a signal
When a slot was disconnected, it was effectively removed
from the list of connections of a Signal. As a result, the
Signal could not notify the connections when it was destroyed.

The fix is that the connection stays in the signal's internal list
until it's destroyed. During emit, we check in the connection itself
if it is connected or not.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>