types: add support for primitive and array types.
The rationale is that primitive and array types are never
put on flow ports. This allows to create smaller typekits
for these data types. Additionally, the fixed size array
typekit has been improved too such that it semantically
creates and manages arrays as the user expects with respect
to copying and initialisation.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>