transport corba : fix serialization of stl container of std::string.
There is a specialization of AnyConversion<std::string>::toCorbaType for TAO implementation to implement the string serialization into a string sequence. This specialization is not choosen in some cases if the element is given by reference. Moreover in tao implementation this element is a kind of smart pointer, so it's intended to be pass by value.