corba: fix losing samples if the corba dispatcher incoming queue overflows
The corba dispatching currently works as follows:

  * a channel queues itself in a internal::List object and triggers
    the corba dispatcher
  * the corba dispatcher, *when woken up*, reads the queue, calls
    transferSamples() and removes the processed channel from the
    queue.

The issue is that, in the lock-free case, internal::List is a
fixed-size list of max. size 20. It is therefore possible to have
the following situation:

  port1 bursts with 30 samples
  port2 writes one sample

if the dispatcher does not get called in-between, the port2 will
not get dispatched and therefore port2 won't get transferred.

This patch does not add a channel element in the queue if it is
already present. It means that the maximum useful list size is
the number of ports in the task. Further work would include making
that size configurable, and/or dynamically sizing it to match the
port count of all bound data flow interfaces.