mqueue: provide compatibility with boost 1.46 or newer
For some reason, we do save the item_version_type, but we never load it,
making the read data corrupt (extra int in input stream). Somehow
we don't play nice with the boost::serialization functions. Can't find
out why, so we just catch the item_version_type and then don't save it.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>